requirements.txt
python-telegram-bot==20.7
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
pytz==2024.1
openpyxl==3.1.2
APScheduler==3.10.4
